Move likeUnlike function to ./models/utilities.js
